Sort tokens more consistently
